pyproject.toml
[tool.poetry]
name = "tutorial-django"
version = "0.1.0"
description = "Tutorial per imparare vari aspetti di Django"
authors = ["Lorenzo Bugli <bugli.lorenzo@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.2.4"
djangorestframework = "^3.14.0"
django-debug-toolbar = "^4.1.0"
drf-yasg = "^1.21.7"
factory-boy = "^3.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
